requirements.txt
pydantic
langchain
langchain-core
langchain-ollama
langchain-community
ddgs
faiss-cpu
sentence-transformers